+++ tests/test_skin_kodi_host.py
@@ -1,5 +1,6 @@
 from __future__ import unicode_literals
 
+import json
 import unittest
 
 from resources.lib.skin_adapter import AF3_ID, skin_settings_document
@@ -23,12 +24,29 @@ class FakeXbmc:
     def __init__(self):
         self.player = FakePlayer()
         self.skin = AF3_ID
+        self.live_settings = []
+        self.appearance = {}
 
     def Player(self):
         return self.player
 
     def getSkinDir(self):
         return self.skin
+
+    def executeJSONRPC(self, document):
+        request = json.loads(document)
+        method = request['method']
+        params = request['params']
+        if method == 'Settings.GetSkinSettings':
+            result = {'skin': self.skin, 'settings': self.live_settings}
+        elif method == 'Settings.GetSettingValue':
+            result = {'value': self.appearance[params['setting']]}
+        elif method == 'Settings.SetSettingValue':
+            self.appearance[params['setting']] = params['value']
+            result = True
+        else:
+            return json.dumps({'id': 1, 'error': {'code': -32601}})
+        return json.dumps({'id': 1, 'result': result})
 
 
 class FakeAddon:
@@ -117,6 +135,7 @@ class KodiSkinHostTests(unittest.TestCase):
             {'id': 'demo.text', 'type': 'string', 'value': 'value'},
         ]
         self.document = skin_settings_document(self.values)
+        self.xbmc.live_settings = list(self.values)
 
     def test_dependency_check_requires_skin_and_helper_versions(self):
         self.host.ensure_dependencies(AF3_ID, 'script.skinvariables')
@@ -165,6 +184,38 @@ class KodiSkinHostTests(unittest.TestCase):
         with self.assertRaisesRegex(KodiSkinHostError, 'inactive'):
             self.host.stage_settings(AF3_ID, self.document, self.values)
         self.assertEqual({}, self.vfs.files)
+
+    def test_activate_skin_uses_switch_callback_and_checks_result(self):
+        self.xbmc.skin = 'skin.estuary'
+        self.host.activate_skin(AF3_ID)
+        self.assertEqual(AF3_ID, self.host.active_skin())
+        self.assertEqual([AF3_ID], self.switches)
+
+        self.xbmc.skin = 'skin.estuary'
+        self.host.switch_skin = lambda _skin: None
+        with self.assertRaisesRegex(KodiSkinHostError, 'did not become'):
+            self.host.activate_skin(AF3_ID)
+
+    def test_verify_loaded_settings_requires_disk_and_live_equality(self):
+        path = self.host._settings_path(AF3_ID)
+        self.vfs.files[path] = self.document
+        self.host.verify_loaded_settings(AF3_ID, self.values)
+
+        self.xbmc.live_settings = []
+        with self.assertRaisesRegex(KodiSkinHostError, 'did not load'):
+            self.host.verify_loaded_settings(AF3_ID, self.values)
+
+    def test_apply_appearance_sets_only_changed_and_verifies(self):
+        self.xbmc.appearance = {
+            'lookandfeel.skintheme': 'SKINDEFAULT',
+            'lookandfeel.skincolors': 'Light',
+        }
+        self.host.apply_appearance({
+            'lookandfeel.skintheme': 'SKINDEFAULT',
+            'lookandfeel.skincolors': 'Dark',
+        })
+        self.assertEqual('Dark', self.xbmc.appearance[
+            'lookandfeel.skincolors'])
 
 
 if __name__ == '__main__':